dag/requirements.txt
psycopg2-binary
pandas
elasticsearch